[font.c] 使用 Integer；去除多余的 extern

+++ fonts.h
@@ -25,8 +25,8 @@ static_assert(UMAXOF(FontIndex) >= FONT_MEM_SIZE,
 extern FourQuarters charinfo(InternalFontNumber f, EightBits p);
 extern Scaled charwidth(InternalFontNumber x, FourQuarters y);
 extern Scaled charitalic(InternalFontNumber x, FourQuarters y);
-extern Scaled charheight(InternalFontNumber x, int y);
-extern Scaled chardepth(InternalFontNumber x, int y);
+extern Scaled charheight(InternalFontNumber x, Integer y);
+extern Scaled chardepth(InternalFontNumber x, Integer y);
 extern Scaled charkern(InternalFontNumber x, FourQuarters y);
 extern FourQuarters get_fontcheck(InternalFontNumber x);
 extern void fonts_init(void);
@@ -35,9 +35,9 @@ readfontinfo(Pointer u, StrNumber nom, StrNumber aire, Scaled s);
 extern int fonts_undump(FILE* fmtfile, FILE* _not_use_);
 extern void fonts_dump(FILE* fmtfile);
 extern long get_hyphenchar(InternalFontNumber x); // texmac.h
-extern void set_hyphenchar(InternalFontNumber x, long c);
+extern void set_hyphenchar(InternalFontNumber x, Integer c);
 extern long get_skewchar(InternalFontNumber x);
-extern void set_skewchar(InternalFontNumber x, long c);
+extern void set_skewchar(InternalFontNumber x, Integer c);
 
 extern FontIndex fmemptr;
 extern MemoryWord fontinfo[FONT_MEM_SIZE + 1];
@@ -46,11 +46,11 @@ extern EightBits fontbc[FONT_MAX + 1];
 extern EightBits fontec[FONT_MAX + 1];
 extern Pointer fontglue[FONT_MAX + 1];
 extern FontIndex bcharlabel[FONT_MAX + 1];
-extern long fontbchar[FONT_MAX + 1];
-extern long fontfalsebchar[FONT_MAX + 1];
-extern long ligkernbase[FONT_MAX + 1]; // texmac
-extern long extenbase[FONT_MAX + 1];
-extern long parambase[FONT_MAX + 1];
+extern Integer fontbchar[FONT_MAX + 1];
+extern Integer fontfalsebchar[FONT_MAX + 1];
+extern Integer ligkernbase[FONT_MAX + 1]; // texmac
+extern Integer extenbase[FONT_MAX + 1];
+extern Integer parambase[FONT_MAX + 1];
 
 // [tex], fonts
 extern StrNumber fontidtext(InternalFontNumber x);
